test: consolidate defaulted_{primitives,enums,collections} into one integration suite
