+++ .github/workflows/integration_tests.yml
@@ -62,19 +62,9 @@ jobs:
           cd integration_test/additional_properties/additional_properties_test
           dart test --concurrency=2
 
-      - name: Run defaulted primitives integration tests
+      - name: Run defaulted integration tests
         run: |
-          cd integration_test/defaulted_primitives/defaulted_primitives_test
-          dart test --concurrency=2
-
-      - name: Run defaulted enums integration tests
-        run: |
-          cd integration_test/defaulted_enums/defaulted_enums_test
-          dart test --concurrency=2
-
-      - name: Run defaulted collections integration tests
-        run: |
-          cd integration_test/defaulted_collections/defaulted_collections_test
+          cd integration_test/defaulted/defaulted_test
           dart test --concurrency=2
 
       - name: Run petstore integration tests
