--- integration_test/defaulted_collections/openapi.yaml
@@ -1,55 +1,0 @@
-openapi: 3.0.4
-info:
-  title: Defaulted Collections API
-  description: >-
-    Schema with `default` values for collection-typed (list / map) and
-    free-form (AnyModel) properties — exercises const-default materialisation
-    for lists, maps, and arbitrary JSON.
-  version: 1.0.0
-  contact:
-    name: Example
-    url: https://example.com
-    email: example@example.com
-
-servers:
-  - url: https://api.example.com/v1
-
-paths:
-  /health:
-    get:
-      operationId: healthCheck
-      summary: Health check
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: string
-
-components:
-  schemas:
-    Filters:
-      type: object
-      properties:
-        tags:
-          type: array
-          items:
-            type: string
-          default: ["new", "featured"]
-        counts:
-          type: object
-          additionalProperties:
-            type: integer
-          default:
-            x: 1
-            y: 2
-        raw:
-          type: object
-          additionalProperties: true
-          default:
-            any: "value"
-            nested: [1, 2, 3]
