--- integration_test/defaulted_primitives/defaulted_primitives_test/analysis_options.yaml
@@ -1,6 +1,0 @@
-include: package:very_good_analysis/analysis_options.yaml
-
-linter:
-  rules:
-    public_member_api_docs: false
-    avoid_print: false
