--- integration_test/defaulted_primitives/defaulted_primitives_test/pubspec.yaml
@@ -1,22 +1,0 @@
-name: defaulted_primitives_test
-description: Integration tests for class-property primitive default values.
-version: 1.0.0
-publish_to: none
-
-environment:
-  sdk: '>=3.10.0 <4.0.0'
-
-dependencies:
-  defaulted_primitives_api:
-    path: ../defaulted_primitives_api
-  dio: ^5.8.0
-  path: ^1.9.1
-  tonik_util: ^0.4.1
-
-dev_dependencies:
-  test: ^1.25.15
-  very_good_analysis: ^10.2.0
-
-dependency_overrides:
-  tonik_util:
-    path: ../../../packages/tonik_util
