+++ pubspec.yaml
@@ -58,19 +58,9 @@ melos:
         cd integration_test/additional_properties/additional_properties_test
         dart test
 
-    test-integration-defaulted-primitives:
+    test-integration-defaulted:
       run: |
-        cd integration_test/defaulted_primitives/defaulted_primitives_test
-        dart test
-
-    test-integration-defaulted-enums:
-      run: |
-        cd integration_test/defaulted_enums/defaulted_enums_test
-        dart test
-
-    test-integration-defaulted-collections:
-      run: |
-        cd integration_test/defaulted_collections/defaulted_collections_test
+        cd integration_test/defaulted/defaulted_test
         dart test
 
     test-integration-petstore:
@@ -247,9 +237,7 @@ melos:
       run: |
         set -e
         melos run test-integration-additional-properties
-        melos run test-integration-defaulted-primitives
-        melos run test-integration-defaulted-enums
-        melos run test-integration-defaulted-collections
+        melos run test-integration-defaulted
         melos run test-integration-petstore
         melos run test-integration-petstore-config
         melos run test-integration-music-streaming
